Change thread model
Using one thread per connection (and also a thread pool) leads to memory corruption because of libmagic, which is not thread-safe!
Current fix (which is not guaranteed...) is to use microhttpd internal select.
Another possible fix is to drop libmagic for mime type in http server and to rely only on file extensions.